Moved popup script, changed namespace and using SignalBus to call ModelController method
