+++ Assets/MenuUi/Scripts/DefenceScreen/CharacterGallery/ModelController.cs
@@ -6,6 +6,7 @@ using Altzone.Scripts.Config;
 using Altzone.Scripts.Model.Poco.Game;
 using Altzone.Scripts.Model.Poco.Player;
 using UnityEngine;
+using SignalBus = MenuUi.Scripts.Lobby.SignalBus;
 
 
 namespace MenuUi.Scripts.CharacterGallery
@@ -20,6 +21,7 @@ namespace MenuUi.Scripts.CharacterGallery
         private void Awake()
         {
             ServerManager.OnLogInStatusChanged += StartLoading;
+            SignalBus.OnRandomSelectedCharactersRequested += SetRandomSelectedCharactersToEmptySlots;
         }
