+++ Assets/MenuUi/Scripts/Lobby/SelectedCharactersPopup.cs
@@ -1,14 +1,25 @@
+using MenuUI.Scripts;
 using MenuUI.Scripts.Lobby.InLobby;
 using UnityEngine;
 
 
-namespace MenuUi.Scripts.CharacterGallery
+namespace MenuUi.Scripts.Lobby
 {
+    public static partial class SignalBus
+    {
+        public delegate void RandomSelectedCharactersRequested();
+        public static event RandomSelectedCharactersRequested OnRandomSelectedCharactersRequested;
+        public static void OnRandomSelectedCharactersRequestedSignal()
+        {
+            OnRandomSelectedCharactersRequested?.Invoke();
+        }
+    }
+
+
     public class SelectedCharactersPopup : MonoBehaviour
     {
         [SerializeField] private GameObject _container;
         [SerializeField] private InLobbyController _inLobbyController;
-        [SerializeField] private ModelController _modelController;
 
         public void OpenPopup()
         {
@@ -24,7 +35,7 @@ namespace MenuUi.Scripts.CharacterGallery
 
         public void SelectRandomCharacters()
         {
-            _modelController.SetRandomSelectedCharactersToEmptySlots();
+            SignalBus.OnRandomSelectedCharactersRequestedSignal();
             ClosePopup();
             _inLobbyController.ToggleWindow();
         }
